+++ apps/cli/src/legacy/commands/functions/list/list.format.ts
@@ -1,0 +1,30 @@
+import { renderGlamourTable } from "../../../output/legacy-glamour-table.ts";
+import type { Functions } from "./list.encoders.ts";
+
+export function formatUnixMilliTimestamp(value: number): string {
+  const date = new Date(value);
+  const parts = [
+    date.getUTCFullYear(),
+    date.getUTCMonth() + 1,
+    date.getUTCDate(),
+    date.getUTCHours(),
+    date.getUTCMinutes(),
+    date.getUTCSeconds(),
+  ];
+  const [year, ...rest] = parts.map((part) => part.toString().padStart(2, "0"));
+  return `${year}-${rest[0]}-${rest[1]} ${rest[2]}:${rest[3]}:${rest[4]}`;
+}
+
+export function renderFunctionsTable(functions: Functions): string {
+  return renderGlamourTable(
+    ["ID", "NAME", "SLUG", "STATUS", "VERSION", "UPDATED_AT (UTC)"],
+    functions.map((fn) => [
+      fn.id,
+      fn.name,
+      fn.slug,
+      fn.status,
+      String(fn.version),
+      formatUnixMilliTimestamp(fn.updated_at),
+    ]),
+  );
+}
